requirements.txt
streamlit
pdfplumber
python-dotenv
langchain-core
langchain-openai
tavily-python


